feat(math): fix and improve average/min/max-Property()
averageProperty(), maxProperty() and minProperty() now return null if the array is empty, instead of throwing an error, and start with the value of the first element instead of 0. They also ensure the property exists and has a correct type for every element of the array. The typing for max and min has been altered to accept Dates and strings as well as numbers.

BREAKING CHANGE: averageProperty(), maxProperty() and minProperty() now return null if the array is empty, instead of throwing an error; and throw an error if the property does not exist or is incorrectly typed for any element of the array.